backend/requirements.txt
# KrishiRaksha AI — FastAPI Backend Requirements
# Install with: pip install -r requirements.txt

# --- Web Framework ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
websockets==12.0

# --- Database ---
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.20.0

# --- Validation ---
pydantic-settings==2.2.1
email-validator==2.1.1

# --- Auth ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- Async HTTP ---
aiohttp==3.9.5

# --- Task Queue ---
celery==5.4.0
redis==5.0.4
flower==2.0.1

# --- Storage ---
boto3==1.34.101
minio==7.2.7

# --- AI / ML ---
openai==1.30.1
google-generativeai==0.5.3
langchain>=0.2.1,<0.3.0
langchain-community>=0.2.1,<0.3.0
langchain-google-genai>=1.0.3,<2.0.0
qdrant-client>=1.9.1,<2.0.0
sentence-transformers==3.0.0

# --- Computer Vision ---
Pillow==10.3.0
opencv-python-headless==4.9.0.80
ultralytics==8.2.0  # Classification training and inference
# torch==2.3.0        # Large download, uncomment when GPU available
# xgboost==2.0.3      # Uncomment for real risk scoring

# --- Utilities ---
numpy<2
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.3.0
pytz==2024.1
phonenumbers==8.13.36

# --- Dev / Debug ---
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
debugpy==1.8.1
rich==13.7.1